Fix Makefile test rule bug
